feat: 전체 클래스 리팩토링 (#9)
