+++ src/main/java/com/example/userserver/auth/util/CookieUtils.java
@@ -58,4 +58,17 @@ public class CookieUtils {
         }
         return null;
     }
+
+    public static String extractRefreshTokenFromCookies(HttpServletRequest request) {
+        Cookie[] cookies = request.getCookies();
+
+        if (cookies != null) {
+            for (Cookie cookie : cookies) {
+                if ("sns-refresh".equals(cookie.getName())) {
+                    return cookie.getValue();
+                }
+            }
+        }
+        return null;
+    }
 }
