+++ src/main/java/com/example/userserver/user/dto/UserRequest.java
@@ -1,10 +1,14 @@
 package com.example.userserver.user.dto;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @Getter
 @Setter
+@NoArgsConstructor
+@AllArgsConstructor
 public class UserRequest {
     private String username;
     private String email;
