+++ src/main/java/com/example/userserver/user/repository/UserRepository.java
@@ -10,5 +10,5 @@ import java.util.Optional;
 public interface UserRepository extends JpaRepository<User, Integer> {
 
     Optional<User> findByUsername(String username);
-    User findByEmail(String email);
+    Optional<User> findByEmail(String email);
 }
